Fix wrong TCP/TLS source port in Contact & Via (#4627)
It was reported that the Contact and Via headers in REGISTER requests are using the listener port instead of the actual port, which did not occur in older PJSIP versions.

Upon investigation, this behavior appears to be introduced by #4411, which always initializes the TCP/TLS local name using the listener's local name, regardless of whether the published address setting is configured

Thanks to Marcus Froeschl for the report.